requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.4
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
packaging==23.1
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
zipp==3.16.2